refactor(ingestion): split ingestion service into app and server modules
